refactor(modules): relocate coding-repos into src/modules/coding/repos
