+++ src/plugins/coding-secrets-facade.ts
@@ -13,7 +13,7 @@ import {
   resolveProviderHost,
 } from '../coding-credentials/resolve-agent-secrets.js'
 import { resolveMcpServers, resolveMcpTokens } from '../coding-credentials/resolve-mcp-servers.js'
-import { getRepoByName, listRepos } from '../coding-repos/store.js'
+import { getRepoByName, listRepos } from '../modules/coding/repos/store.js'
 import { deny } from './deny.js'
 import type { CodingRepoEntry } from './runtime-types.js'
 import type { PluginToolRuntimeContext } from './types.js'
